feat: proxy requests to graph
