+++ src/api/config.ts
@@ -1,0 +1,2 @@
+export const graphUrl =
+  process.env.GRAPHQL_URL ?? "https://dev-bag-of-holding.herokuapp.com/";
