+++ src/api/gql/index.ts
@@ -1,3 +1,2 @@
 export { LoginGQL } from "./loginMutation";
-export { MeGQL } from "./meQuery";
 export * from "./__generated__";
