--- src/config.ts
@@ -1,5 +1,0 @@
-export const config = {
-  graphURL:
-    process.env.NEXT_PUBLIC_GRAPHQL_URL ??
-    "https://dev-bag-of-holding.herokuapp.com/",
-};
