+++ src/pages/api/graphql.ts
@@ -1,0 +1,8 @@
+export { graphqlProxy as default } from "api/graphql";
+
+export const config = {
+  api: {
+    bodyParser: false,
+    externalResolver: true,
+  },
+};
