+++ src/profile/gql/index.ts
@@ -1,0 +1,2 @@
+export { MeGQL } from "./meQuery";
+export * from "./__generated__/Me";
